feat: Add dev container configuration for developing inside a container
This commit adds a devcontainer.json file that allows developers to use
a docker file to develop inside a container. By configuring the dev
container, anyone with docker installed can easily develop in this
environment. The devcontainer.json file includes settings for building
the container, forwarding the port 8080, and customizing VSCode with
specific extensions. This feature greatly simplifies the setup process
for developers. close #6